test: rename and organize tests
